Make prompt a child of the preview view
So that it closes when the preview closes, along with any associated
notifications.
