+++ lib/glsl-preview-view.js
@@ -10,6 +10,7 @@ const interact = require("interactjs");
 const ShaderCanvas = require("shader-canvas");
 
 const Config = require("./config");
+const Prompt = require("./prompt");
 
 function getActiveTab() {
   return document.querySelector(".tab[data-type=\"GlslPreviewView\"]");
@@ -73,6 +74,8 @@ module.exports = class GlslPreviewView {
     this.element.appendChild(this.shader.domElement);
     this.disposables.add(this.shader);
 
+    this.prompt = new Prompt();
+
     this.userSize = {};
     interact(this.shader.domElement).resizable({
       edges: {left: true, right: true, bottom: true, top: true},
@@ -116,6 +119,8 @@ module.exports = class GlslPreviewView {
   destroy() {
     if (this.IS_DESTROYED) { return; }
     this.IS_DESTROYED = true;
+
+    this.prompt.destroy();
 
     this.disposables.dispose();
     this.element.innerHTML = "";
@@ -353,11 +358,12 @@ module.exports = class GlslPreviewView {
 
       encoder.run().then(() => {
         const elapsedSeconds = (performance.now() / 1000) - startTimeSeconds;
-        atom.notifications.addSuccess("Video saved", {
+        const n = atom.notifications.addSuccess("Video saved", {
           detail: `Saved to: ${outputPath} in ${elapsedSeconds}
           seconds (${elapsedSeconds / numFrames} seconds/frame)`,
           dismissable: true,
         }); // TODO: dispose?
+        this.disposables.add(new Disposable(n.dismiss.bind(n)));
       }).error((e) => {
         const opts = {
           detail: e,
@@ -366,7 +372,8 @@ module.exports = class GlslPreviewView {
         if (e.message === "spawn ffmpeg ENOENT") {
           opts.detail = "ffmpeg not found, is it installed?";
         }
-        atom.notifications.addError("Video recording failed", opts); // TODO: dispose?
+        const n = atom.notifications.addError("Video recording failed", opts);
+        this.disposables.add(new Disposable(n.dismiss.bind(n)));
       }).finally(() => {
         if (!pausedAtStart) {
           // We could unpause this as soon as we're finished recording, but I kind
@@ -376,6 +383,24 @@ module.exports = class GlslPreviewView {
           this.shader.togglePause();
         }
       });
+    });
+  }
+
+  saveVideoStream() {
+    const defaultDuration = "2";
+    const message = "Enter video duration in seconds";
+    this.prompt.open(message, defaultDuration).then((value) => {
+      const durationString = value || defaultDuration;
+      const duration = parseFloat(durationString);
+      if (isNaN(duration)) {
+        const n = atom.notifications.addError("Unable to parse duration", {
+          detail: "Duration should be a float in seconds.",
+          dismissable: true,
+        });
+        this.disposables.add(new Disposable(n.dismiss.bind(n)));
+        return;
+      }
+      this.recordVideoStream(duration, 60);
     });
   }
 };
