+++ lib/main.js
@@ -2,7 +2,6 @@ const url = require("url");
 const {CompositeDisposable, Disposable} = require("atom");
 
 const Config = require("./config");
-const Prompt = require("./prompt");
 
 let GlslPreviewView = null;
 
@@ -61,8 +60,6 @@ module.exports = {
     }));
 
     this.disposables.add(atom.workspace.addOpener(opener));
-
-    this.prompt = new Prompt();
   },
 
   toggle() {
@@ -135,21 +132,7 @@ module.exports = {
   saveVideoStream() {
     const preview = this.getActivePreview();
     if (!preview) { return; } // TODO: show notification
-    const defaultDuration = "2";
-    const message = "Enter video duration in seconds";
-    this.prompt.open(message, defaultDuration).then((value) => {
-      const durationString = value || defaultDuration;
-      const duration = parseFloat(durationString);
-      if (isNaN(duration)) {
-        const n = atom.notifications.addError("Unable to parse duration", {
-          detail: "Duration should be a float in seconds.",
-          dismissable: true,
-        });
-        this.disposables.add(new Disposable(n.dismiss.bind(n)));
-        return;
-      }
-      preview.recordVideoStream(duration, 60);
-    });
+    preview.saveVideoStream();
   },
 
   upgradeCheck() {
@@ -167,8 +150,6 @@ As always, feedback and bug reports are much appreciated!`,
 
   deactivate() {
     this.disposables.dispose();
-
-    this.prompt.destroy();
 
     atom.workspace.getPaneItems().forEach((item) => {
       if (isGlslPreviewView(item)) {
